Extracts AuthenticationMode Descriptor
